+++ CHANGES.md
@@ -1,5 +1,9 @@
 ### Changelogs
 
+##### 1.0.2
++ add sd.util.Romanize with test spec (using scalatest)
++ Romanize username on result sms `SmsResult.OkCharged`
+
 ##### 1.0.1
 Change SmsResult.OkValid's json field `status` from 0 to 1.
 As documented at http://developers.1pay.vn/http-apis/smsplus-charging: